libfreerdp-sspi: avoid conflicting definitions when NATIVE_SSPI is on

+++ libfreerdp-sspi/NTLM/ntlm_message.c
@@ -540,7 +540,7 @@ SECURITY_STATUS ntlm_write_ChallengeMessage(NTLM_CONTEXT* context, SecBuffer* bu
 
 	if (context->NegotiateFlags & NTLMSSP_REQUEST_TARGET)
 	{
-		TargetNameLen = context->TargetName.cbBuffer;
+		TargetNameLen = (uint16) context->TargetName.cbBuffer;
 		TargetNameBuffer = context->TargetName.pvBuffer;
 	}
 	else
@@ -553,7 +553,7 @@ SECURITY_STATUS ntlm_write_ChallengeMessage(NTLM_CONTEXT* context, SecBuffer* bu
 
 	if (context->NegotiateFlags & NTLMSSP_NEGOTIATE_TARGET_INFO)
 	{
-		TargetInfoLen = context->TargetInfo.cbBuffer;
+		TargetInfoLen = (uint16) context->TargetInfo.cbBuffer;
 		TargetInfoBuffer = context->TargetInfo.pvBuffer;
 	}
 	else
@@ -941,14 +941,14 @@ SECURITY_STATUS ntlm_write_AuthenticateMessage(NTLM_CONTEXT* context, SecBuffer*
 	if (context->ntlm_v2 < 1)
 		WorkstationLen = 0;
 
-	DomainNameLen = context->identity.DomainLength;
+	DomainNameLen = (uint16) context->identity.DomainLength;
 	DomainNameBuffer = (uint8*) context->identity.Domain;
 
-	UserNameLen = context->identity.UserLength;
+	UserNameLen = (uint16) context->identity.UserLength;
 	UserNameBuffer = (uint8*) context->identity.User;
 
-	LmChallengeResponseLen = context->LmChallengeResponse.cbBuffer;
-	NtChallengeResponseLen = context->NtChallengeResponse.cbBuffer;
+	LmChallengeResponseLen = (uint16) context->LmChallengeResponse.cbBuffer;
+	NtChallengeResponseLen = (uint16) context->NtChallengeResponse.cbBuffer;
 
 	EncryptedRandomSessionKeyLen = 16;
 	EncryptedRandomSessionKeyBuffer = context->EncryptedRandomSessionKey;
